requirements.txt
DateTime==4.1.1
Django==1.10.1
djangorestframework==3.6.2
inflect==0.2.5
logger==1.4
mysqlclient==1.3.10
pdfkit==0.6.1
pytz==2017.2
requests==2.13.0
validate-email==1.3
wkhtmltopdf==0.2
xlrd==1.0.0
zope.interface==4.3.3
